Feat/subtask

+++ backend/app/models/requirement.py
@@ -5,8 +5,8 @@ from app.models.application import Application
 class Requirement(db.Model):
     requirement_id = db.Column(db.Integer, primary_key=True)
     tenant_id = db.Column(db.Integer, nullable=False)
-    requirement_name = db.Column(db.String(255), nullable=False)
-    original_requirement = db.Column(db.String(1000))
+    requirement_name = db.Column(db.String(6000), nullable=False)
+    original_requirement = db.Column(db.String(6000))
     app_id = db.Column(db.Integer, nullable=False)
     username = db.Column(db.String(100))
     default_source_branch = db.Column(db.String(255))
@@ -14,8 +14,8 @@ class Requirement(db.Model):
     status = db.Column(db.String(20))
     satisfaction_rating = db.Column(db.Integer)
     completion_rating = db.Column(db.Integer)
-    created_at = db.Column(db.TIMESTAMP, default=db.func.current_timestamp())
-    updated_at = db.Column(db.TIMESTAMP, default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
+    created_at = db.Column(db.String(100), default=db.func.current_timestamp())
+    updated_at = db.Column(db.String(100), default=db.func.current_timestamp(), onupdate=db.func.current_timestamp())
 
     @staticmethod
     def create_requirement(tenant_id, requirement_name, original_requirement, app_id, username, default_source_branch, default_target_branch, status, satisfaction_rating=None, completion_rating=None):
@@ -36,11 +36,11 @@ class Requirement(db.Model):
         return requirement
 
     @staticmethod
-    def get_all_requirements(tenantID=None):
-        requirements = Requirement.query.filter_by(tenant_id=tenantID).order_by(Requirement.requirement_id.desc()).all()
+    def get_all_requirements(tenantID=None, page=1, per_page=40):
+        requirements = Requirement.query.filter_by(tenant_id=tenantID).order_by(Requirement.requirement_id.desc()).paginate(page=page, per_page=per_page, error_out=False)
         requirement_list = []
 
-        for req in requirements:
+        for req in requirements.items:
             req_dict = {
                 'requirement_id': req.requirement_id,
                 'requirement_name': req.requirement_name,
@@ -57,10 +57,16 @@ class Requirement(db.Model):
             }
             requirement_list.append(req_dict)
 
-        return requirement_list
+        return {
+            'requirements': requirement_list,
+            'total_pages': requirements.pages,
+            'current_page': requirements.page,
+            'total_items': requirements.total
+        }
 
     @staticmethod
-    def get_requirement_by_id(requirement_id):
+    def get_requirement_by_id(requirement_id, tenant_id=0):
+        tenant_id = int(tenant_id)
         req = Requirement.query.get(requirement_id)
         if req:
             req_dict = {
@@ -68,22 +74,29 @@ class Requirement(db.Model):
                     'requirement_name': req.requirement_name,
                     'original_requirement': req.original_requirement,
                     'app_id': req.app_id,
+                    'tenant_id': req.tenant_id,
                     'username': req.username,
                     'default_source_branch': req.default_source_branch,
                     'default_target_branch': req.default_target_branch,
                     'status': req.status,
                     'satisfaction_rating': req.satisfaction_rating,
                     'completion_rating': req.completion_rating,
-                    'created_at': req.created_at,
-                    'updated_at': req.updated_at,
+                    'created_at': str(req.created_at),
+                    'updated_at': str(req.updated_at),
                     'app': Application.get_application_by_id(req.app_id)
                 }
+            if tenant_id and tenant_id != req_dict["tenant_id"]:
+                return None
             return req_dict
         return None
 
     @staticmethod
-    def update_requirement(requirement_id, **kwargs):
+    def update_requirement(requirement_id, tenant_id, **kwargs):
+        tenant_id = int(tenant_id)
         requirement = Requirement.query.get(requirement_id)
+        if tenant_id and tenant_id != requirement.tenant_id:
+            return None
+        
         if requirement:
             for key, value in kwargs.items():
                 setattr(requirement, key, value)
@@ -93,8 +106,11 @@ class Requirement(db.Model):
         return None
 
     @staticmethod
-    def delete_requirement(requirement_id):
+    def delete_requirement(requirement_id, tenant_id):
+        tenant_id = int(tenant_id)
         requirement = Requirement.query.get(requirement_id)
+        if tenant_id and tenant_id != requirement.tenant_id:
+            return None
         if requirement:
             db.session.delete(requirement)
             db.session.commit()
